Ayuba_Voting_DApp/backend/requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
whitenoise==6.0.0
dj-database-url==0.5.0
psycopg2==2.9.3
